requirements.txt
langchain==0.3.1
langchain-groq==0.2.1
langchain-community==0.3.1
wikipedia==1.4.0
python-dotenv==1.0.1